requirements.txt
livekit-agents>=1.3.0
livekit-plugins-groq
livekit-plugins-sarvam
livekit-plugins-silero
python-dotenv